third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/xsl-blocked-expected.txt

CONSOLE ERROR: Loading the script 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/style.xsl' violates the following Content Security Policy directive: "script-src 'unsafe-inline'". Note that 'script-src-elem' was not explicitly set, so 'script-src' is used as a fallback. The action has been blocked.
CONSOLE WARNING: crbug.com/435623334: This page uses XSLT, which being considered for removal from the web. If that happens, it is possible that this page will need to be updated to maintain functionality.
